order 프리팹설정 완

+++ multigame/Assets/Script/Order.cs
@@ -7,11 +7,11 @@ public class Order
   private FoodData food;
   public FoodData Food{ get { return food; } }
 
-  private int timeLimit;
-  public int TimeLimit{ get{ return timeLimit; } }
+  private float timeLimit;
+  public float TimeLimit { get{ return timeLimit; } }
 
 
-  public Order(FoodData food,int timeLimit){
+  public Order(FoodData food,float timeLimit){
     this.food = food;
     this.timeLimit = timeLimit;
   }
